requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.6
soundfile==0.12.1
librosa==0.10.0
numpy==1.24.3
llama-cpp-python==0.2.85
chromadb==1.4.0
tiktoken==0.12.0
redis==7.0.1